add more safeguards to release.sh

+++ release.sh
@@ -10,8 +10,12 @@ if [[ "$V" == "" ]] ; then
     echo "error: \$V not set"
 else
     [[ -e "$ZIP" ]] && rm "$ZIP"
-    [[ -f "$DEST" ]] || mkdir "$DEST"
-    rm -rf "${DEST:?}/*" # avoid rm /* if $DEST is empty. thanks shellcheck!
-    for f in $FILES ; do cp -a "$f" "$DEST" ; done
-    zip -r "$ZIP" "$DEST"
+    if [[ -f "$DEST" ]] ; then
+        echo "error: '$DEST' exists and is not a folder"
+    else
+        [[ -d "$DEST" ]] || mkdir "$DEST"
+        rm -rf "${DEST:?}/*" # avoid rm /* if $DEST is empty. thanks shellcheck!
+        for f in $FILES ; do cp -a "$f" "$DEST" ; done
+        zip -r "$ZIP" "$DEST"
+    fi
 fi
